requirements.txt
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
lpips>=0.1.4
torchmetrics>=1.2.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.14.0
pandas>=2.0.0
scikit-image>=0.21.0

